+++ examples/state.rs
@@ -74,7 +74,7 @@ fn main() {
     HttpServer::new(
         Application::builder("/", AppState{counter: Cell::new(0)})
             // enable logger
-            .middleware(Logger::new(None))
+            .middleware(middlewares::Logger::new(None))
             // websocket route
             .resource("/ws/", |r| r.get::<MyWebSocket>())
             // register simple handler, handle all methods